requirements.txt
PyYAML
pytest
redis
# Pycrypto has dependencies 'python3-devel' and 'redhat-rpm-config' on the Fedora RHEL build machine
pycrypto
